+++ app/src/main/java/nl/mpcjanssen/simpletask/Constants.kt
@@ -61,6 +61,7 @@ object Constants {
     const val EXTRA_PREFILL_TEXT = "prefill_text"
     const val EXTRA_TASK_ID = "taskid"
     const val EXTRA_TARGET_TODO_FILE = "target_todo_file"
+    const val EXTRA_FROM_LAUNCHER_SHORTCUT = "from_launcher_shortcut"
 
     // Android OS specific constants
     const val ANDROID_EVENT = "vnd.android.cursor.item/event"